feat: implement identity module with authentication and token management
