+++ src/producer/application/service/index.ts
@@ -1,2 +1,1 @@
 export * from './producer-application.service';
-export * from './producer-auth.service';
