+++ src/producer/domain/model/index.ts
@@ -4,4 +4,9 @@ export * from './crop';
 export * from './farm';
 export * from './harvest';
 export * from './producer';
+export * from './password';
+export * from './document-validator.interface';
 export * from './document-validator.factory';
+export * from './encrypter.interface';
+export * from './encrypt-password';
+export * from './password.factory';
