+++ src/producer/web/rest/dto/request/index.ts
@@ -1,2 +1,1 @@
 export * from './create-producer.request';
-export * from './login-producer.request';
